default now accepts a callback
Anything that responds to :call can be used as a default value.  Such
defaults will be evaluated on instantiation, but after all non-callable
defaults have been evaluated (so that they can be used for any callable
default).  Callable defaults are run in the order in which they appear
in the model definition.

This resolves #1.